+++ examples/scripts/spm_MLE.py
@@ -57,7 +57,7 @@ x, final_cost = optim.run()
 print("Estimated parameters:", x)
 
 # Plot the timeseries output
-pybop.quick_plot(problem, inputs=x[0:2], title="Optimised Comparison")
+pybop.quick_plot(problem, problem_inputs=x[0:2], title="Optimised Comparison")
 
 # Plot convergence
 pybop.plot_convergence(optim)
@@ -69,5 +69,5 @@ pybop.plot_parameters(optim)
 pybop.plot2d(likelihood, steps=15)
 
 # Plot the cost landscape with optimisation path
-bounds = np.array([[0.55, 0.77], [0.48, 0.68]])
+bounds = np.asarray([[0.55, 0.77], [0.48, 0.68]])
 pybop.plot2d(optim, bounds=bounds, steps=15)