+++ pybop/parameters/parameter.py
@@ -1,9 +1,11 @@
 from collections import OrderedDict
-from typing import Dict, List
+from typing import Dict, List, Union
 
 import numpy as np
 
 from pybop._utils import is_numeric
+
+Inputs = Dict[str, float]
 
 
 class Parameter:
@@ -169,7 +171,7 @@ class Parameters:
         for param in args:
             self.add(param)
 
-    def __getitem__(self, key: str):
+    def __getitem__(self, key: str) -> Parameter:
         """
         Return the parameter dictionary corresponding to a particular key.
 
@@ -426,7 +428,7 @@ class Parameters:
                 values = self.true_value()
         return {key: values[i] for i, key in enumerate(self.param.keys())}
 
-    def verify(self, inputs=None):
+    def verify(self, inputs: Union[Inputs, None] = None):
         """
         Verify that the inputs are an Inputs dictionary or numeric values
         which can be used to construct an Inputs dictionary