+++ pybop/problems/fitting_problem.py
@@ -2,8 +2,7 @@ import numpy as np
 
 from pybop import BaseProblem
 from pybop._dataset import Dataset
-from pybop.models.base_model import Inputs
-from pybop.parameters.parameter import Parameters
+from pybop.parameters.parameter import Inputs, Parameters
 
 
 class FittingProblem(BaseProblem):
@@ -94,13 +93,13 @@ class FittingProblem(BaseProblem):
         inputs = self.parameters.verify(inputs)
 
         requires_rebuild = False
-        for key in inputs.keys():
-            if (
-                key in self._model.rebuild_parameters
-                and inputs[key] != self.parameters[key].value
-            ):
-                self.parameters[key].update(value=inputs[key])
-                requires_rebuild = True
+        for key, value in inputs.items():
+            if key in self._model.rebuild_parameters:
+                current_value = self.parameters[key].value
+                if value != current_value:
+                    self.parameters[key].update(value=value)
+                    requires_rebuild = True
+
         if requires_rebuild:
             self._model.rebuild(parameters=self.parameters)