+++ tests/integration/test_spm_parameterisations.py
@@ -11,7 +11,7 @@ class Test_SPM_Parameterisation:
 
     @pytest.fixture(autouse=True)
     def setup(self):
-        self.ground_truth = np.array([0.55, 0.55]) + np.random.normal(
+        self.ground_truth = np.asarray([0.55, 0.55]) + np.random.normal(
             loc=0.0, scale=0.05, size=2
         )
 
@@ -160,7 +160,7 @@ class Test_SPM_Parameterisation:
         [
             pybop.SciPyDifferentialEvolution,
             pybop.IRPropMin,
-            pybop.CMAES,
+            pybop.XNES,
         ],
     )
     @pytest.mark.integration
@@ -218,7 +218,7 @@ class Test_SPM_Parameterisation:
         cost = pybop.RootMeanSquaredError(problem)
 
         # Select optimiser
-        optimiser = pybop.CMAES
+        optimiser = pybop.XNES
 
         # Build the optimisation problem
         optim = optimiser(cost=cost)
@@ -245,7 +245,5 @@ class Test_SPM_Parameterisation:
             ]
             * 2
         )
-        sim = model.predict(
-            init_soc=init_soc, experiment=experiment, inputs=parameters.as_dict(x)
-        )
+        sim = model.predict(init_soc=init_soc, experiment=experiment, inputs=x)
         return sim