+++ tests/unit/test_plots.py
@@ -89,7 +89,7 @@ class TestPlots:
         pybop.quick_plot(design_problem)
 
         # Test conversion of values into inputs
-        pybop.quick_plot(fitting_problem, inputs=[0.6, 0.6])
+        pybop.quick_plot(fitting_problem, problem_inputs=[0.6, 0.6])
 
     @pytest.fixture
     def cost(self, fitting_problem):
@@ -127,6 +127,12 @@ class TestPlots:
 
         # Plot the cost landscape with optimisation path
         pybop.plot2d(optim, steps=5)
+
+        # Plot the cost landscape using optimisation path
+        pybop.plot2d(optim, steps=5, use_optim_log=True)
+
+        # Plot gradient cost landscape
+        pybop.plot2d(optim, gradient=True, steps=5)
 
     @pytest.mark.unit
     def test_with_ipykernel(self, dataset, cost, optim):